pyproject.toml
[project]
name = "deepsite-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncclick>=8.1.8",
    "google-adk>=0.3.0",
    "httpx>=0.28.1",
]
